requirements.txt
Flask==3.0.3
python-dotenv>=0.19.2
authlib>=1.0
flask-sqlalchemy==3.1.1
Werkzeug==2.3.7
mindee==3.11.0
requests==2.31.0
sendgrid==6.10.0
qrcode==7.4.2
Pillow==10.0.1
web3==6.10.0
eth-account==0.9.0
boto3==1.28.53
email-validator==2.0.0
gunicorn==21.2.0
Flask-Migrate==4.0.5
PyJWT==2.8.0
cryptography==41.0.4
pyotp==2.9.0
Flask-WTF==1.2.1